requirements.txt
Django
coinbase-commerce
django-crispy-forms
django-allauth
django-taggit
django-tinymce-filebrowser
django-filebrowser-no-grappelli
bootstrap4
django-bootstrap4
django-bootstrap4-form
django-haystack
django-analytical
django-progressbarupload